# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Astral Modulator Public API
  version: 2.0.0-scaffold
  license:
    name: MIT
    identifier: MIT
  description: |
    人类与多 LLM Agent 的协作控制面。CLI（astral-cli 仓库）与 Web GUI 共用本协议。

    v2（容器化任务树，TODO.md D15）：workspace 与 task 都是容器，子任务集合
    统一为 GET/POST <container>/children；全量平铺列表与 /tasks/search 已移除，
    平面查询走 /workspaces/{id}/task-search。protocol_version 升为 2，
    与 v1 不兼容（破坏性变更，开发期不做兼容窗口）。

    已裁决的文档分歧（详见 TODO.md「文档分歧裁决」D1-D4）：
    - Device flow 采用 `/auth/device/authorizations*` 路径。
    - 当前身份端点为 `/auth/me`。

    通用约定：
    - 所有时间戳为 RFC3339 UTC。
    - ID 为 `prefix_uuidv7` 形式的不透明字符串，客户端不得解析内部结构。
    - 错误响应统一使用 ErrorEnvelope；CLI 依赖稳定 error.code。
    - 乐观并发统一使用 body 内 expected_revision（protocol §6 方案 A）。
servers:
  - url: /api/v1
    description: 路径前缀固定 /api/v1（不随协议大版本变化）；协议大版本以
      /.well-known/astral 的 protocol_version 为准
...
security:
  - bearerAuth: []
paths:
...
  /workspaces/{workspace_id}/presence/me:
    put:
      tags: [presence]
      summary: Presence 心跳（展示状态，不代表任务所有权）
      operationId: putPresence
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PresenceInput" }
      responses:
        "200":
          description: 已记录
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Presence" }
  /workspaces/{workspace_id}/presence:
    get:
      tags: [presence]
      summary: workspace 在线状态总览（GUI 首页数据源）
      operationId: listPresence
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
      responses:
        "200":
          description: presence 列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Presence" }
  /workspaces/{workspace_id}/messages:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
    post:
      tags: [messages]
      summary: 发送消息（支持 Idempotency-Key，重试不得双发）
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MessageSend" }
      responses:
        "201":
          description: 已发送
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
    get:
      tags: [messages]
      summary: 消息列表（按 target/thread 过滤）
      operationId: listMessages
      parameters:
        - { name: thread_id, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: 消息分页列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePage"
...
  /tasks/{task_id}/messages:
    get:
      tags: [messages]
      summary: task thread 消息
      operationId: listTaskMessages
      parameters:
        - { $ref: "#/components/parameters/TaskId" }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: 消息分页列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePage"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Human access token 或 Agent/Service credential（不共用）
    cookieAuth:
      type: apiKey
      in: cookie
      name: astral_session
      description: Web GUI HttpOnly+Secure+SameSite session（与 CLI Bearer 同 actor 不同 session）
  parameters:
    WorkspaceId:
      name: workspace_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
    TaskId:
      name: task_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string, description: 不透明游标，取自 next_cursor }
...
  schemas:
    Id:
      type: string
      pattern: "^[a-z]{2,3}_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      description: 不透明 ID：prefix_uuidv7。前缀 usr/agt/svc/srv/ws/tsk/tgp/tag/doc/cfl/prs/msg/evt/cred/ses/dev/req/aud/apv/inv。
      examples: ["ws_0192ab34-56cd-7ef8-9a01-234567890abc"]
    IdOrNull:
      oneOf:
        - $ref: "#/components/schemas/Id"
        - type: "null"
...
    PresenceState:
      type: string
      enum: [idle, planning, working, waiting, blocked, reviewing]
    PresenceInput:
      type: object
      required: [state]
      properties:
        state: { $ref: "#/components/schemas/PresenceState" }
        current_task_id: { $ref: "#/components/schemas/Id" }
        note: { type: string, maxLength: 500 }
        ttl_seconds: { type: integer, minimum: 30, maximum: 300, default: 90 }
    Presence:
      type: object
      required: [actor_id, state, last_heartbeat_at]
      properties:
        actor_id: { $ref: "#/components/schemas/Id" }
        display_name: { type: string }
        state: { type: string, description: 含派生值 offline（由 expires_at 判定） }
        current_task_id: { $ref: "#/components/schemas/Id" }
        note: { type: string }
        last_heartbeat_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
    MessageSend:
      type: object
      required: [target, body]
      properties:
        target:
          type: object
          required: [type, id]
          properties:
            type: { type: string, enum: [actor, workspace, task] }
            id: { $ref: "#/components/schemas/Id" }
        thread_id: { $ref: "#/components/schemas/IdOrNull" }
        body: { type: string, minLength: 1, maxLength: 20000 }
        metadata: { type: object, additionalProperties: true }
    Message:
      type: object
      required: [id, workspace_id, sender_id, body, created_at]
      properties:
        id: { $ref: "#/components/schemas/Id" }
        workspace_id: { $ref: "#/components/schemas/Id" }
        thread_id: { $ref: "#/components/schemas/IdOrNull" }
        sender_id: { $ref: "#/components/schemas/Id" }
        body: { type: string }
        metadata: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
...
    PageMeta:
      type: object
      required: [next_cursor]
      properties:
        next_cursor:
          type: [string, "null"]
          description: null 表示没有更多
...
    MessagePage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/Message" } }
...